voxtts/pyproject.toml
[project]
name = "voxtts"
version = "0.1.0"
description = "Text-to-Speech with local GPU/CPU engines for natural voice synthesis"
requires-python = ">=3.11"
dependencies = [
    "kokoro>=0.9",
    "soundfile>=0.12",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "lameenc>=1.7.0",
    "lingua-language-detector>=2.0",
    "pip",  # required by spaCy (kokoro dependency) at runtime
]

[project.optional-dependencies]
piper = ["piper-tts>=1.2.0"]

[project.scripts]
voxtts = "voxtts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voxtts"]
